requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt
psycopg2-binary
django-cors-headers
drf-spectacular
django-filter
python-decouple==3.8
Pillow
gunicorn